Relocate data fetch and update page title
Relocate the data fetch code to the page level instead of dashboard
component (originally to pass `repoName`, which didn't work, but this
location seems like a cleaner location for it).

Update the page title to reflect the project/package name, not just
ORCA.

Integrate `lang` attribute into `<html>` tag.
